feat: add max pages, move types
